requirements.txt
fastapi
uvicorn
supabase
pyjwt
passlib
sqlalchemy
psycopg2
pydantic
python-dotenv